Allow MythArchive to handle very large (>4Gb) MythVideo files.
Fixes #4431.


git-svn-id: http://svn.mythtv.org/svn/trunk@15345 7dbf422c-18fa-0310-86e9-fd20926502f2

+++ mythplugins/mytharchive/mytharchive/videoselector.h
@@ -13,14 +13,14 @@
 
 typedef struct
 {
-    int id;
+    int     id;
     QString title;
     QString plot;
     QString category;
     QString filename;
     QString coverfile;
     int     parentalLevel;
-    uint    size;
+    unsigned long long size;
 } VideoInfo;
 
 class VideoSelector : public MythThemedDialog
